finish ma env indicator seeries

+++ lib/src/deriv_chart/indicators_ui/ma_env_indicator/ma_env_indicator_config.dart
@@ -27,11 +27,11 @@ class MAEnvIndicatorConfig extends IndicatorConfig {
   /// Field type
   final String fieldType;
 
-  /// Shift type
+  /// Moving Average Envelope shift type
   final ShiftType shiftType;
 
   /// Moving Average Envelope shift
-  final int shift;
+  final double shift;
 
   /// MA line style
   final LineStyle lineStyle;
